Refactor endpoint trace graph network nodes
